Renamed curve operation precompile files to a simpler/more consistent naming scheme
