+++ lib/index.js
@@ -12,9 +12,9 @@ const num02 = require('./precompiled/02-sha256.js')
 const num03 = require('./precompiled/03-ripemd160.js')
 const num04 = require('./precompiled/04-identity.js')
 const num05 = require('./precompiled/05-modexp.js')
-const num06 = require('./precompiled/06-bn128_add.js')
-const num07 = require('./precompiled/07-bn128_mul.js')
-const num08 = require('./precompiled/08-bn128_pairing.js')
+const num06 = require('./precompiled/06-ecadd.js')
+const num07 = require('./precompiled/07-ecmul.js')
+const num08 = require('./precompiled/08-ecpairing.js')
 
 module.exports = VM
