+++ lib/precompiled/08-ecpairing.js
